Support generics in functions + records
I believe this works, there maybe cases where we supply the args into
the wrong order, but I'm pretty sure this is right. It'd be nice to keep
around the name of the generic (if there is one) but that is a little
annoying with how the type system is structured right now. We probably
need a lexical scope similar to how we're tracking variables.

Also fixed some issues with the prelude - we're able to compile stuff
now :tada: